pyblob/requirements.txt
Flask
azure-identity
azure-storage-blob